refactor: General code refactoring and fixes
Mostly reducing repetitive device tree code

+++ include/hw/arm/apple-silicon/boot.h
@@ -284,7 +284,7 @@ typedef struct {
     hwaddr dram_base;
     uint64_t dram_size;
     uint8_t nvram_data[XNU_MAX_NVRAM_SIZE];
-    uint64_t nvram_size;
+    uint32_t nvram_size;
     char *ticket_data;
     uint64_t ticket_length;
     uint8_t boot_nonce_hash[XNU_BNCH_SIZE];
